Decomposition rebuilt
