+++ encrypt.py
@@ -2,8 +2,6 @@ import logging
 import os
 from cryptography.hazmat.primitives import padding as s_padding
 from cryptography.hazmat.primitives.ciphers import Cipher, algorithms, modes
-from cryptography.hazmat.primitives.asymmetric import padding
-from cryptography.hazmat.primitives import hashes
 
 
 def encrypt_symmetric(path_to_text: str, symmetric_key: bytes, save_path: str) -> None:
@@ -35,16 +33,3 @@ def encrypt_symmetric(path_to_text: str, symmetric_key: bytes, save_path: str) -
     except OSError as err:
         logging.warning(f'{err} during writing to {save_path}')
 
-
-def encrypt_asymmetric(public_key: bytes, key: bytes) -> bytes:
-    """
-    Symmetric key encryption
-    :param public_key:
-    :param key: Key for symmetric encryption
-    :return:
-    """
-    encr_key = public_key.encrypt(key,
-                                  padding.OAEP(mgf=padding.MGF1(algorithm=hashes.SHA256()), algorithm=hashes.SHA256(),
-                                               label=None))
-    logging.info(f'Symmetric key has been encrypted')
-    return encr_key
